Software source is now managed by Nasqueron
Initial plan for Tommy was to let Zooniverse manage the source code.

The dereckson/Tommy fork was a temporary Git repository to
send pull requests to zooniverse, so they can upgrade the code.

Six months later, we can consider Zooniverse doesn't want to maintain
Tommy anymore. As Nasqueron uses it at https://builds.nasqueron.org
it makes sense to maintain it as a Nasqueron adopted project.