Add support for Jenkins 2.0 multi-branch pipelines
Summary:
Jenkins 2.0 added projects with automatic discovery of Jenkinsfile.

This pipeline-as-code syntax is aware of Git branches, and as such,
jobs are created for each branch discovered by Jenkins or configured.

Such projects don't currently appear on the dashboard, as the root
project is only a container object, where actual projects is in the
jobs property.

This change requires a deeper JSON document and parse sub-projects.

Fixes T1141.

Test Plan: Run server for Nasqueron CD

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1141

Differential Revision: https://devcentral.nasqueron.org/D1794